Fixed bug whereby with multiproducers (or in the future, MC) you could read a reference before it was written. This is due to writeAcquire() advancing the producerIndex (or readAcquire advancing the consumerIndex in the MC) case.
When the producerReferenceArrayIndex() (or consumerReferenceArrayIndex)
method is called it reads the current producerIndex. Clearly sometimes
this will have changed between writeAcquire and
producerReferenceArrayIndex calls. Thus P0 who won the writeAcquire,
could end up skipping over it's ref array slot and competing on the
same slot as P1.

The result is nulls in some slots and incorrect values in others.